refactor(order-management): extract pre-sign hearing check into OrderSignValidator strategy
Move validateHearingNotAlreadyScheduled out of BSSService into a generic
OrderSignValidator strategy (ScheduledHearingSignValidator), dispatched by
OrderSignValidationService which fetches the orders once and runs every
validator. New pre-sign rules now plug in as @Component beans, matching the
existing OrderUpdateStrategy/OrderStrategyExecutor convention.
