+++ backend/dristi-services/order-management/src/main/java/pucar/service/BSSService.java
@@ -38,10 +38,10 @@ public class BSSService {
     private final OrderServiceFactoryProvider factoryProvider;
     private final ADiaryUtil aDiaryUtil;
     private final HearingUtil hearingUtil;
-    private final CompositeOrderService compositeOrderService;
+    private final OrderSignValidationService orderSignValidationService;
 
     @Autowired
-    public BSSService(XmlRequestGenerator xmlRequestGenerator, ESignUtil eSignUtil, FileStoreUtil fileStoreUtil, CipherUtil cipherUtil, OrderUtil orderUtil, Configuration configuration, OrderServiceFactoryProvider factoryProvider, ADiaryUtil aDiaryUtil, HearingUtil hearingUtil, CompositeOrderService compositeOrderService) {
+    public BSSService(XmlRequestGenerator xmlRequestGenerator, ESignUtil eSignUtil, FileStoreUtil fileStoreUtil, CipherUtil cipherUtil, OrderUtil orderUtil, Configuration configuration, OrderServiceFactoryProvider factoryProvider, ADiaryUtil aDiaryUtil, HearingUtil hearingUtil, OrderSignValidationService orderSignValidationService) {
         this.xmlRequestGenerator = xmlRequestGenerator;
         this.eSignUtil = eSignUtil;
         this.fileStoreUtil = fileStoreUtil;
@@ -51,7 +51,7 @@ public class BSSService {
         this.factoryProvider = factoryProvider;
         this.aDiaryUtil = aDiaryUtil;
         this.hearingUtil = hearingUtil;
-        this.compositeOrderService = compositeOrderService;
+        this.orderSignValidationService = orderSignValidationService;
     }
 
     public List<OrderToSign> createOrderToSignRequest(OrdersToSignRequest request) {
@@ -60,9 +60,9 @@ public class BSSService {
         log.info("creating order to sign request, result= IN_PROGRESS, orderCriteria:{}", request.getCriteria().size());
 
         // Pre Validation
-        // Orders that create a new hearing (SCHEDULE_OF_HEARING_DATE / SCHEDULING_NEXT_HEARING, including
-        // such items inside a composite order) cannot be published if the case already has a scheduled hearing.
-        validateHearingNotAlreadyScheduled(request);
+        // Runs all pre-sign validators (e.g. a case cannot be published with a new hearing order if it
+        // already has a scheduled hearing). New rules plug in via the OrderSignValidator strategy.
+        orderSignValidationService.validate(request);
 
         List<CoordinateCriteria> coordinateCriteria = new ArrayList<>();
 
@@ -113,75 +113,6 @@ public class BSSService {
         return orderToSign;
 
     }
-
-    /**
-     * Validates that none of the orders selected for (bulk) signing will create a new hearing for a case
-     * that already has a hearing in SCHEDULED status. If such a conflict is found a CustomException is thrown
-     * so the user is informed before the orders are signed/published.
-     */
-    private void validateHearingNotAlreadyScheduled(OrdersToSignRequest request) {
-        RequestInfo requestInfo = request.getRequestInfo();
-
-        Set<String> conflictingFilingNumbers = new LinkedHashSet<>();
-        for (OrdersCriteria criterion : request.getCriteria()) {
-            OrderSearchRequest searchRequest = OrderSearchRequest.builder()
-                    .requestInfo(requestInfo)
-                    .criteria(OrderCriteria.builder()
-                            .orderNumber(criterion.getOrderNumber())
-                            .tenantId(criterion.getTenantId()).build())
-                    .build();
-
-            OrderListResponse orders = orderUtil.getOrders(searchRequest);
-            if (orders.getList() == null || orders.getList().isEmpty()) {
-                continue;
-            }
-
-            Order order = orders.getList().get(0);
-            if (!createsHearing(order)) {
-                continue;
-            }
-
-            if (hasScheduledHearing(requestInfo, order)) {
-                log.error("Hearing already scheduled for case with filingNumber:{}, orderNumber:{}", order.getFilingNumber(), order.getOrderNumber());
-                conflictingFilingNumbers.add(order.getFilingNumber());
-            }
-        }
-
-        if (!conflictingFilingNumbers.isEmpty()) {
-            String filingNumbers = String.join(", ", conflictingFilingNumbers);
-            throw new CustomException(HEARING_ALREADY_SCHEDULED_ERROR, "A hearing is already scheduled for the following case(s): " + filingNumbers);
-        }
-    }
-
-    private boolean createsHearing(Order order) {
-        // An order carrying a nextHearingDate schedules a follow-up hearing on sign
-        // (BSSService.updateOrderWithSignDoc -> HearingUtil.preProcessScheduleNextHearing).
-        if (order.getNextHearingDate() != null) {
-            return true;
-        }
-        if (COMPOSITE.equalsIgnoreCase(order.getOrderCategory())) {
-            return compositeOrderService.getItemListFormCompositeItem(order).stream()
-                    .anyMatch(item -> isHearingCreatingOrderType(item.getOrderType()));
-        }
-        return isHearingCreatingOrderType(order.getOrderType());
-    }
-
-    private boolean isHearingCreatingOrderType(String orderType) {
-        return SCHEDULE_OF_HEARING_DATE.equalsIgnoreCase(orderType) || SCHEDULING_NEXT_HEARING.equalsIgnoreCase(orderType);
-    }
-
-    private boolean hasScheduledHearing(RequestInfo requestInfo, Order order) {
-        List<Hearing> hearings = hearingUtil.fetchHearing(HearingSearchRequest.builder()
-                .requestInfo(requestInfo)
-                .criteria(HearingCriteria.builder()
-                        .filingNumber(order.getFilingNumber())
-                        .tenantId(order.getTenantId()).build())
-                .build());
-
-        return Optional.ofNullable(hearings).orElse(Collections.emptyList()).stream()
-                .anyMatch(hearing -> SCHEDULED.equalsIgnoreCase(hearing.getStatus()));
-    }
-
 
     private String generateRequest(String base64Doc, String timeStamp, String txnId, String coordination, String pageNumber) {
         log.info("generating request, result= IN_PROGRESS, timeStamp:{}, txnId:{}, coordination:{}, pageNumber:{}", timeStamp, txnId, coordination, pageNumber);
